+++ src/drivers/driver_unsupported.c
@@ -12,6 +12,10 @@ bool driver_is_unknown() {
   return false;
 }
 
+bool driver_is_link_state_unknown() {
+  return true;
+}
+
 void driver_get_system_mac(uint8_t mac[ETH_HWADDR_LEN]) {
   LWIP_UNUSED_ARG(mac);
 }